refactor

+++ notification-service/Program.cs
@@ -1,8 +1,9 @@
-using FirebaseAdmin;
-using Google.Apis.Auth.OAuth2;
+using MassTransit;
 using Microsoft.AspNetCore.Mvc;
-using notification_service.Config;
+using Microsoft.EntityFrameworkCore;
+using notification_service.dbContext;
 using notification_service.DTOs;
+using notification_service.Entities;
 using notification_service.Interfaces;
 using notification_service.Repositories;
 using notification_service.Services;
@@ -12,14 +13,44 @@ var builder = WebApplication.CreateBuilder(args);
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
 
-builder.Services.Configure<MongoDbSettings>(
-    builder.Configuration.GetSection("MongoDb:NotificationsDatabase"));
+builder.Services.AddDbContext<AppDbContext>(options =>
+    options.UseNpgsql(
+        "Host=localhost;Port=5430;Username=admin;Password=admin;Database=notification-service-db"
+    )
+);
 
-builder.Services.AddSingleton<INotificationRepository, MongoNotificationRepository>();
-builder.Services.AddSingleton<INotificationService, NotificationService>();
+builder.Services.AddScoped<INotificationRepository, PostgresNotificationRepository>();
+builder.Services.AddScoped<INotificationService, NotificationService>();
 //builder.Services.AddSingleton<INotificationSender, FirebaseNotificationSender>();
 
+builder.Services.AddMassTransit(config =>
+{
+    config.SetKebabCaseEndpointNameFormatter();
+    config.AddConsumers(typeof(Program).Assembly);
+    config.UsingRabbitMq((context, cfg) =>
+    {
+        cfg.Host(new Uri("amqp://localhost:5672"), h =>
+        {
+            h.Username("guest");
+            h.Password("guest");
+        });
+        cfg.ConfigureEndpoints(context);
+    });
+});
+
 var app = builder.Build();
+
+using (var scope = app.Services.CreateScope())
+{
+    var dbContext = scope.ServiceProvider.GetRequiredService<AppDbContext>();
+    var pendingMigrations = dbContext.Database.GetPendingMigrations();
+    if (pendingMigrations.Any())
+    {
+        Console.WriteLine("Applying pending migrations...");
+        dbContext.Database.Migrate();
+        Console.WriteLine("Migrations applied successfully.");
+    }
+}
 
 if (app.Environment.IsDevelopment())
 {
@@ -65,6 +96,14 @@ app.MapGet("/api/notifications/user/{userid}", async (
     }
 });
 
+app.MapPost("/api/notifications/", async (
+    INotificationService notificationService, 
+    [FromBody] CreateNotificationDto notificationDto) =>
+{
+    var response = await notificationService.CreateNotificationAsync(notificationDto);
+    return Results.Ok(response);
+});
+
 app.MapPut("/api/notifications/status", async (
     INotificationService notificationService, 
     [FromBody] UpdateNotificationStatusDto updateDto) =>
@@ -86,8 +125,10 @@ app.MapDelete("/api/notifications/{id}", async (
 {
     try
     {
-        var result = await notificationService.DeleteNotificationAsync(id);
-        return !result ? Results.NotFound() : Results.NoContent();
+        var result = await notificationService.GetNotificationByIdAsync(id);
+        if(result == null) return Results.NotFound();
+        await notificationService.DeleteNotificationAsync(id);
+        return Results.NoContent();
     }
     catch (Exception ex)
     {
